+++ client/src/components/volunteersPage/VolunteersPage.tsx
@@ -10,11 +10,6 @@ const VolunteersPage = () => {
   const { backend } = useBackendContext();
   const [totalVolunteers, setTotalVolunteers] = useState(0);
   const [totalHours, setTotalHours] = useState(0);
-  const [refreshStatus, setRefreshStatus] = useState(true);
-
-  // const [volunteers, setVolunteers] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState(null);
 
   useEffect(() => {
     const fetchData = async () => {
@@ -26,19 +21,12 @@ const VolunteersPage = () => {
         setTotalHours(totalHoursResponse.data.totalHours);
         setTotalVolunteers(totalVolunteersResponse.data.totalVolunteers);
       } catch (err) {
-        setError(err.message);
-      } finally {
-        setLoading(false);
+        console.error(err);
       }
-    };
-
-    if (refreshStatus) {
-      fetchData();
-      setRefreshStatus(false);
     }
 
     fetchData();
-  }, [backend, refreshStatus]);
+  }, [backend]);
 
 
   return (
